Make combinators available at jail.combinators
Instead of requiring the third argument to be a function that passes in
combinators, this exposes all combinators on `jail.combinators` and
allows the third argument to just be a list (or even null!). This may be
more ergonomic in some contexts (like nixpkgs overlays).

This isn't a breaking change, the third argument can still be a function
and will receive the combinators if it is.