Fix missing `''` in add-cleanup docs
This one is a bit confusing since these docs document a nix multi-line
string within a multi-line string, *and* there is a bash interpolation
within. We want the docs to render as `if [ -e "''${TMPFILE-}" ]`, since
the docs are within a multi-line string.

Sorry in advance for anyone attempting to parse this in the nix source
code instead of the docs website :(.