Update readonly-paths-from-var test case
In 8fa6d132e16f03b9d1997fe6d446634c104f5485, ~mathiassven submitted a
patch that fixes the `readonly-paths-from-var` combinator omitting the
final path.

The tests didn't catch this since the final path in the test was a
non-existent path.

This commit backfills the test to catch regressions related to this bug.
The test now makes sure that the first, last, and at least one middle
test path are real paths that show up in the jail.