Forward `override` to jailed derivations
On Nixos, it may be desirable to install some software using
`programs.*.enable = true`. Jail.nix mostly supports this because you
can either use `jail-nix.mkOverlay`, or just use Nixos
`programs.*.package = jail ...` to inject a jailed version of the
package.

However, for some modules, like Firefox, the Nixos module calls
`override` on the package. Normally, you would just call `override` on
the package passed into jail, but this is not possible to do within a
Nixos module.

To address this, this patch will check to see if a package passed into
jail has an `override` field, and if so will add an `override` on the
jailed derivation that passes the override function to the entrypoint.

Thanks to ~juanjakobo for reporting this!

Fixes https://todo.sr.ht/~alexdavid/jail.nix/1