Fix `try-fwd-env` setting unset variables
If an environment variable is unset, `try-fwd-env` will set it to an
empty string. This can cause issues for programs that treat unset and
empty environment variables differently. For instance, when jailing
Spotify with the `pulse` combinator, `$PULSE_SERVER` gets set to an
empty string, Spotify tries to use it as-is, fails, and then falls back
to using ALSA.

Ideally I would've liked to wire this into the existing env logic
somehow but it's a bit tricky to do since `--setenv` needs to be
provided to Bubblewrap conditionally based on a value that isn't known
until the script is executed. Bubblewrap appears to handle `--setenv`
being specified multiple times for the same variable okay, so not using
the existing env logic doesn't break the ability to override environment
variables.